test(session): stabilize legacy stderr pid kill fixture
Keep daemon-like test process evidence stable across platforms by moving the
shared daemon fixture into a small support module and isolating the legacy
stderr PID kill coverage outside the existing monolith test file.

This preserves the product guard that refuses stale or reused PIDs while making
the cross-platform fixture match the liveness evidence expected by production
kill validation on macOS.

### AI Reviewer Metadata
- **Design Intent**: Fix PR #2248 macOS all-features CI without weakening session kill PID reuse protection.
- **Key Decisions**: Keep shell as the session leader, embed the session id in the command line, poll liveness recognition before kill, and split the new coverage out of the monolith test file.
- **Reviewer Guidance**:
  - **Timing/Race Scenarios**: macOS `ps` can expose child command lines differently from Linux `/proc`; the fixture must prove the PID is recognized as live session evidence before kill.
  - **Boundary Cases**: stale or reused PIDs must still be rejected by product code.
  - **Regression Tests Added**: session_cmds_daemon::session_cmds_daemon_kill_tests::handle_session_kill_accepts_legacy_stderr_pid
